Cleanup our Docker dev env.
Move the services onto a common network shared by all other apps.
Don't remove volumes on restart and rebuild. The dev can do that if
they need to b/c they want to nuke everything. We should rely on a
generally working env and not require nuking the volumes in the normal
dev flow. Use COPY instead of ADD b/c ADD wasn't doing what I thought it
was. COPY is recommended and volumes are mounted through the
docker-compose.yml file so that changes to files inside and outside the
container are reflected in both places.