Fix up docker dev env so that changes made outside the container are reflected inside.
This is done by mounting the entire codebase as a volume when started
using docker-compose.

Note: a dependency of webmock (dev builds only) got updated to the point
of only supporting ruby 2.3 and above. So I locked it down to the last
version that supported ruby 2.1. The dependency tree was:

  webmock was resolved to 1.24.6, which depends on
    addressable was resolved to 2.7.0, which depends on
      public_suffix was resolved to 4.0.1

Also, stop supporting a local dev. All dev should be done with the
container. Removed Gemfile.lock for the docker build so that everything
inside is the same as outside and it's like you're just developing
locally with the Docker container acting as the web-server.