Make Docker restarts/rebuilds work when new gem versions are available.
We weren't .dockerignore'ing Gemfile.lock, so when you last built
the Docker image it built whatever Gemfile.lock (which is .gitignored)
into the image. This may not match what `bundle install` installs and things
fail. Just ignore this file in the image build for now. The proper solution
is to check-in Gemfile.lock and have a process for how to update it, but since
we're planning on getting rid of this app/server, I'm not investing time to fix
this up properly.

TESTING:
- Did a fresh docker build (no cache) and it failed to start. Made this change
  and rebuilt. It started.