Staging and prod run Ruby 1.9.3. A prior gem I pinned was for Ruby 2.1.
Pinning it at the last version supported for Ruby 1.9.3. We had
never done a staging of prod release after the last time I changed this
(or just generally for years and years). So this make this repo at
least deployable to staging and prod even if none of the recent changes
have needed to be deployed b/c they are all for dev env stuff.

TESTING:
- manually made the change on staging and ran `bundle install --path vendor/bundle`
- restarted the server and it worked