Dev env more like staging/prod + checked in Gemfile.lock
We should have our Gemfile.lock checked in so that we can
understand the behavior / differences when bundle install
uses new versions of gems. But, to do this we needed the dev
env to more closely match staging / prod, hence downgrading Ruby
to 1.9.3. I verified that the Gemfile.lock generated here matches
staging/prod.

Note: originally I built this dev env out with Ruby 2.1 b/c the
plan was to start updating / patching all our apps to be on the same
version of Ruby and it just seemed to work. But now we're getting rid of
this web app, so in the meantime, I just want the dev env to work properly
and when things change/fail have them be easy to diagnose / fix.

TESTING:
- rebuilt the container (including blowing away the DB volume) and was
  able to login.